[INFRA/CORE] Move logger to pkg
